requirements.txt
click
moviepy
Pillow
numpy
face_recognition
